pyproject.toml
[project]
name = "timelineparser"
version = "0.1.0"
description = "Google Timeline Export Parser — parse Google Maps Timeline JSON exports (Google Takeout) into SQLite"
requires-python = ">=3.11"
dependencies = [
    "tqdm",
 "requests",
 "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
timelineparser = "timelineparser.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["ruff", "pytest", "mypy", "types-requests", "types-tqdm"]

[tool.ruff]
line-length = 88

[tool.mypy]
ignore_missing_imports = true
